// js_modules/dagster-ui/packages/ui-core/src/assets/types/AssetNodeOverview.types.ts
export type AssetOverviewMetadataEventsQueryVariables = Types.Exact<{
  assetKey: Types.AssetKeyInput;
}>;